+++ gazebo/physics/ode/ODEBallJoint.hh
@@ -30,6 +30,9 @@ namespace gazebo
 {
   namespace physics
   {
+    /// \addtogroup gazebo_physics_ode
+    /// \{
+
     /// \brief An ODEBallJoint
     class GZ_PHYSICS_VISIBLE ODEBallJoint : public BallJoint<ODEJoint>
     {
@@ -81,6 +84,7 @@ namespace gazebo
       // Documentation inherited.
       protected: virtual void SetForceImpl(unsigned int _index, double _torque);
     };
+    /// \}
   }
 }
 #endif